+++ facade/src/test/scala/react/semanticui/addons/portal/PortalInnerTests.scala
@@ -1,0 +1,17 @@
+package react.semanticui.addons.portal
+
+import utest._
+import japgolly.scalajs.react.test._
+import japgolly.scalajs.react.vdom.html_<^._
+
+object PortalInnerTests extends TestSuite {
+  val tests = Tests {
+    test("render") {
+      val portal = PortalInner(child = <.div("Abc"))
+      ReactTestUtils.withNewBodyElement { mountNode =>
+        portal.renderIntoDOM(mountNode)
+        assert(mountNode.outerHTML == """<div></div>""")
+      }
+    }
+  }
+}